gateway: add /welcomer channel command. fix welcome images not being sent when images are enabled. handle when no channel is configured.
Add /welcomer text command and handler. /welcomer test will now return messages when all modules are disabled or a channel is not configured. Added new error when a channel is not configured and welcomer images or welcomer text is enabled. If welcomer DMs are enabled, this error is not returned. Fixed images not being sent when welcomer text is not enabled. Added optimization to welcomer-gateway to not fetch user from state cache if DMs are disabled, as it is only used for user DMs. Refreshed dependencies